abs-ai-hub/apps/rag-pdf-voice/transcriber/requirements.txt
fastapi==0.111.0
uvicorn==0.30.0
openai-whisper==20240930
torch==2.1.0
torchaudio==2.1.0
